requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.11
djangorestframework==3.15.1
sqlparse==0.4.4
typing-extensions==4.10.0
tzdata==2024.1
